Remove mailhost setting from smokeping
It was not used anyway because Net::SMTP perl package was not installed
on the server and the corresponding "if" was failing. Thus it was failing
back to using the binary configured in sendmail (in pathnames file). Do
note however that even if that is not configured it will again fallback
to using /usr/lib/sendmail which is a valid link to exim4 in our
environment

Change-Id: I9d67f60e11045558edb337c89e5f9ed52c7f9b11