contacts.wm - remove ferm/443, remove SSL cert
this role moved behind misc-web a while back.
the backend is zirconium. during that move we
already removed the SSL part from the Apache config
as well. so we also should _not_:

- install an SSL cert
- open firewall on 443

because the backend only speaks http now

also to delete the cert in RT #8630 and I7fe03827332d

Change-Id: I040f09e8523c9934071c219e39eefb045b22c383
RT: 8630
RT: 8560