puppetmaster's pid ensured absent
puppetmaster's init script does not clean the pid. Not only that but
it also uses it to report on the status of the puppetmaster service.
This creates a race condition that makes it possible for the
puppetmaster service to be forcefully stopped again. This seems to
interfere sometimes with passenger which starts to spew 500s

Change-Id: I087cb6473c43bf4b3abf81f14e24d7e019c1108f