trebuchet: derive the grain name from the repo name
It's a bit silly to have to configure a grain name for each Trebuchet-deployed
repository, and then have to look up the repository in the configuration to
derive the key. It buys you nothing except the ability to have a single grain
associated with multiple repositories, which isn't beneficial anyway: if the
repositories are part of some role, then they should be declared together in
the role class, rather than have their association declared in a giant hash
blob in manifests/role/deployment.pp.

Change-Id: I38443f3b582c4658f91f0f8d442a590a27aa5b68