ishmael - remove SSL config remnants
This used to be a standalone Apache with
an SSL config but we moved it behind misc-web
and the Apache SSL config part is already gone
for a while.

So this means these settings in the role and
installing the certificate is pointless now.

Remove this, so we can also delete the cert itself.

RT: 8654
RT: 6732

Change-Id: Ib3f83dc1cf160464da91c07ec33629659fb659bc
reference: I8dd78b7fc57c146
reference: I653e1b303d76f5e3392