Require 2 ACKs from kafka brokers for mobile caches
With the default of only 1 ACK, we got missing cache log lines when a
broker dropped out of it's partition leader role.

Previous tests showed that requiring of 2 ACKs per batch solved the
issue in 4 of 4 instances. So it seems 2 ACKs are sufficient, and we
don't need to require ACKs of all ISR machines.

While the expected increase in network traffic should not affect
network performance, we're only slowly ramping up the number of
machines that request more ACKs.

For now it's only the mobile cluster. Other clusters will follow, if
this goes as expected.

Bug: 69667
Change-Id: I241984b375ce65c95cafd76dd3bb22bdd0aa71f7