role::labs::instance: include sudo::labs_project
At the moment, we provision 3 default classes to the labs nodes, one of
which is sudo::labs_project; it doesn't make sense (and it makes it
harder to find) not to include this rather tiny class explicitly in role::labs::instance.
After this is merged, I'll remove the class from the wikitech config
where it's actually included.

Change-Id: I22571a240d5e754bb24e8809db569dd7aa6b68af
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>