hhvm: add a 'creates' clause for source code installation
The HHVM package is ensure => present, which means that Puppet won't be
provisioning it, except on the initial run. Thus having the source code
installation exec resource subscribe to the package means it will never get run
for already-provisioned hosts. Change it to a creates => instead. This does
have the downside of requiring us to explicitly rm -rf /usr/local/src/hhvm any
time we want the source code updated.

Change-Id: I8f16b61f2580292d928864d786d6910091bb98bf