logstash: Remove upstart jobs
The logstash packages produced by logstash.net include both System
V init scripts and upstart jobs for the logstash services. Since we
specify `provider => 'debian'` we are telling Puppet to prefer to
execute the System V scripts. This is in conflict with the default
behavior of `service logstash start` which will prefer to execute the
upstart scripts. This leads to duplicate service execution on servers
where an administrator has manually started the service.

Bug: 72202
Change-Id: Id6b9101e594d051098fa66c66e3c10d69d494588