Make BounceHandler extension work on en-wiki
Due to convenience and security reasons, we are installing BounceHandler
on en-wiki to which we are expecting maximum bounces to happen. This
change affects polonium and will make sure that bounces targeted at
en-wikipedia.org will route through the mwbounceprocessor router and
transport to get HTTP POSTED to the extension installed in en-wiki.

Added en.wikipedia.org to the +local_domains because mwbounceprocessor router
wants the $doamain_part of the return path address in the +local_domain and
+verp_domain - else - the bounce gets redirected via signed smtp.

Removed +local_domain from the realm switch as it was no longer required there.

Change-Id: I7badb2383a3e4d6d78e28bea6a97c5a51d5be64d