Make BounceHandler extension work on meta-wiki
Due to convenience and database query ease, we are installing the extension
on beta. This change affects polonium and will make sure that bounces targeted at
meta.wikimedia.org will route through the mwbounceprocessor router and
transport to get HTTP POSTED to the extension installed in en-wiki.

Removed +local_domain from the realm switch as it was no longer required there

Change-Id: I7badb2383a3e4d6d78e28bea6a97c5a51d5be64d