RT - add mail aliases
on the RT server, magnesium, exim is configured like
"# Use the system aliasfile /etc/aliases for system domains"
and rt.wikimedia.org is a system domain

in T675 for phabricator, we will want to be able to
redirect just some of the RT queues over to phab
but without redirecting them all

first i wanted to puppetize the entire /etc/aliases
but paravoid pointed out how there is the native puppet
type mailalias for that. so i'm using that.

the "ops-request"->"ops-requests" thing already exists
on the server like that. and the other redirect is to
test forwarding things to phabricator

later we will forward some queues over to phabricator
but not all of them at once

Bug: T675
Change-Id: I0f81033477188dae4821b8b25eb729078a4d83d2