hiera: mediawiki-based backend for labs
This backend will retreive its data from the mediawiki api, and will
cache results absolutely for cache_ttl seconds not to kill the backend,
and while they're not changed (based on the mediawiki revision) in order
to speed up the process.

It will be processed for all labs instances, after the traditional yaml
backend we use for prod as well.

Change-Id: I3710fdb23828693701025aa3879dd4a5813d4088
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>