Elasticsearch java options for gc issues
If there are GC issues you want to know quickly and get as much information
as possible.  This causes GC issues to stop the application and dump the
heap for analysis.  It defines GC issues as full GCs that pause the world
for more then 70% of the time and/or can't push heap usage under 90%.
Change-Id: I85ceba638e0a4f0ed5fad87fb2bc086db5616fac