Add halfak to research group
Aaron is most definitely a researcher, and should have been in this group to
begin with.  We are still waiting for a complete list of researchers, but
we expect this password to change soon, and don't want Aaron to be
without the password, so I am adding him to this group now.

Change-Id: I9ce9e3a616f668e30031636ae0ad07f74b08f12b