lists - disable SSLv3
This will not disable SSLv3 on lists.wikimedia.org for now, because
we are using lighttpd/1.4.26, but disabling SSLv3 was not supported
until 1.4.29. (http://www.lighttpd.net/2011/7/3/1-4-29/)

Nevertheless, I think it's a good idea to update the configuation,
so that when we update the server, it will take effect immediately.

Change-Id: I56282aa31b26f69350cf1743c5b46de3715e98a8