Periodically restart HHVM
Once a day, check the uptime of HHVM. If HHVM has been running for more than a
day, restart it. We do this as a form of insurance against subtle memory leaks.
A graceful restart once a day is nicer than gradual memory exhaustion followed
by a visit from the OOM killer.

Change-Id: I43304d2135a4a20370b83e58c5a29ca0cfaa5c17