syslog-ng: keep 30 days worth of logs
This originated from a request to keep more swift logs in related RT.

Rationale: a day worth of compressed logs is ~1GB and swift is the biggest part
of that. Judging from the puppet recipe the syslog server used to be on NFS
which is no longer the case and the current syslog server has ~300GB to spare.
So 30 days seems like an easy win while still respect the data retention policy
of 90 days.

also style/convetion fix invoke-rc.d vs service

RT: 8781

Change-Id: If0eb2b46e3e709f6f6ebae435177960c745648e2