contint: Zuul status learned to query a single change
The Zuul status page learned to expose the status of a single change
with an URL such as /status/change/X,Y. The daemon on gallium runs
behind an Apache proxy which has a whitelist of URL patterns.

Add a pattern that matches a Gerrit (change, patchset) such as 1234,1
limiting the lenght of the arguments to prevent potentially nasty usage.
The pattern accepts up to a billion change and 999 patchsets which
should be fine.

Bug: 47609
Change-Id: Ic5016371b40457d36acf56054bd6440b5cd5585d