Catch a couple of corner cases in check_puppetrun
time_ago that was introduced in 5a3fe15 did not catch a couple of corner
cases. Notably if s was 0 or 1 the entire sorted reversed struct would be
returned and periods like 60 seconds would not be converted to 1 minute
and so on. Fix this by testing for s == 0, change > to >= and provide a
default return value in case of other failures

Change-Id: Iecaafb4ea696fab68a3ad3e7c857626ef2371cff