mailman: enable rmlist for web-list deletion
After a quick discussion with Daniel on IRC about mailing lists,
I thought it might be a nice idea to put this out for a discussion.

This enables the /mailman/rmlist/$list option for mailing lists.
When a listadmin goes to that page, they are prompted to to input
the list administrator password and then given a 'delete archives
or not' option. If they select no, then the list is no longer
listed on /listinfo and any emails to the list bounce. If the
yes option is selectd, the archives are also delete.

If the archives are not deleted - they are still accessible in the
usual place.

This does not expose any security risks in mailman or the like but
does mean a rogue listadmin is a little worse. The upside is, it
means ops members are not needed to delete lists any more and can
act as a permanent 'list disable'.

Change-Id: Id21e65c4a2434ebe5cab5b4ca06ab021011fc66b