Install maven on analytics clients
Using maven, people no longer need to upload >50MB jars again and
again when developing for the Analytics cluster, but they can build
the jars locally.

The system-wide default is to fetch artifacts through our archiva. But
this is not bullet-proof, as people can override those defaults in
user settings and use http_proxies to fetch from the web directly.

RT: 8765
Change-Id: I97e9e5b8e20afedfa66a2da129aa3cae1b626378