Consolidate debugging-related configurations in hhvm::debug
The various tidbits designed to make debugging HHVM more pleasant were clogging
up init.pp, which I'd like to be a concise description of the things that are
absolutely essential to the setup (rather than peripheral to it, as the
debugging helpers are). For that reason, introduce a separate hhvm::debug
module, and move the debugging symbol package declarations there, too. Since
this leaves hhvm::packages practically empty, move the hhvm package
declarations back to init.pp.

Add a fix to a libstdc++ pretty-printer and source HHVM's pretty printers in
/etc/gdb/gdbinit.

Change-Id: Ie3e5f6cb158e89dd4aaf0fe765bc39e2306ce8e3