Install libvips, libvips-dev and libvips-tools on Trusty.
We had a custom -dev and -tools package set up for Trusty
that depended on an unavailable version of libvips.  I've
removed those packages from the Trusty repo.

This patch also specifies libvips37 for Trusty.  That's the
same thing that we use on production Trusty hosts.  Note that
the -dev and -tools packages automatically pull in libvips37,
so there's no concern with order of operations.

Of course, this change discards whatever customizations may
have been added to the precise libvips-dev and libvips-tools
packages.  I'm hoping it's acceptable to just use the standard
but up-to-date packages here.

Change-Id: I20ec5eebc936c6ce3b5c1f83fda0cdccef46da8e