Add cron job that generates flow statistics
Background
==========
The mobile team has a python script that they use to generate statistics
from the analytics-store database cluster to simple CSV output files.
The cron configured in statistics.pp runs that script and syncs those
files to stat1001 where they're publicly served for dashboard
consumption.

This Change
===========
The flow team would like to do roughly the same thing, so we are
repurposing the mobile team's script but defining configuration for it
in a separate repository so as to not interfere.  We're also running it
in a separate process to prevent failures from one config affecting the
other.

Change-Id: Ic8724d12ea008a3bf803a51160c82f1d191e0cf7