hhvm: ensure that jemalloc heap profiling is disabled.
This means that if you want to capture heap profiles, you have to temporarily
disable Puppet. But that's OK, because you should not have heap profiling
turned on in production for more than a few hours. This way we can be sure
we're not forgetting to turn it off (which results in degraded performance and
junk in /var/tmp/hhvm).

Change-Id: I4f3534ea2aaf67c6ee29267ab9843933b8390de8