swift: report statsd data to localhost
swift statsd traffic seems to account for >= 50% of the statsd traffic, thus
driving txstatsd cpu usage very high on tungsten. A more reasonable solution is
to have statsd locally and let txstatsd push it to tungsten.

NOTE: this works only because the swift metrics are per-machine, and thus don't
need to get aggregated across machines in a central statsd

Change-Id: Ia95564b1b2950d16d44c347e1a612ef5b079822c