Move OCG logging off of root partition; logrotate hourly.
We are using a size-limited logrotate configuration, but the size often
grows far past the bounds in a single day.  Run logrotate hourly in order
to ensure that our size limits are respected.

The OCG machines have a very small root partition; move logging to
the much-larger /srv partition.  Leave a symlink in /var/log/ocg
so that unfamiliar admins can easily find the logs.

Change-Id: Id323dd9795f82b2307aa53f025a97e67e355a727