logstash: Use conditional instead of deprecated grep filter
From startup logging: "The 'grep' plugin is no longer necessary now that
you can do if/elsif/else in logstash configs. This plugin will be
removed in the future. If you need to drop events, please use the drop
filter. If you need to take action based on a match, use an 'if' block
and the mutate filter. See the following URL for details on how to use
if/elsif/else in your logstash configs:
http://logstash.net/docs/1.4.2-modified/configuration"

Replace the deprecated grep filter with a conditional regex match and
a drop filter.

Change-Id: I20c08a6c6c4c7ea391b04377c78badf3551893d7