gerrit: configure sshd to not listen on gerrit IP
on hosts using the gerrit role, we have 2 IP addresses,
one is the host (ytterbium.wikimedia.org) and one is the
service (gerrit.wikimedia.org). both are configured on eth0

as described in Bug 35611 / RT: 8838 what we want is to
enable Gerrit listening on port 22, so that for example
users behind corporate firewalls don't have trouble accessing
the current high port 29418 we are using exclusively

it sounds like the host already has the 2 IPs for this exact
reason, we just never finished this

I743d0993a6273a suggest to make Gerrit listen on 22, but first
we need to make sure SSHd does NOT listen on that IP anymore, instead
just on the server IP. this change is the attempt to make that
happen. we have to puppetize the entire SSHd config because
per default we'd do 0.0.0.0:22

sshd_config manpage says hostnames are OK to use here
ListenAddress host|IPv4_addr|IPv6_addr

so we do that and use the DNS name instead of hardcoding IP

Change-Id: Id27657ca29e41960886b517fe0be63ad992d382a