mediawiki: move beta::mwdeploy_sudo to mediawiki::users
Currently, when deployers run scap, scap SSHs them into each application
servers and runs commands as mwdeploy and as apache by running 'sudo -u
mwdeploy ..', 'sudo -u apache ..', etc.  This is possible because users in the
wikidev group are allowed to sudo to mwdeploy and apache. We're changing scap
to make it SSH into each host as mwdeploy, using the shared keyholder key. To
make the migration smooth, we need to continue to allow scap to run commands on
application servers by running 'sudo -u mwdeploy' or ('-u apache'). At the
moment, however, mwdeploy is not allowed to sudo as itself.

Bryan has already done this for beta, so it's simply a matter of moving the
code to mediawiki::users rather than the beta-specific beta::mwdeploy_sudo.

This scheme conforms to the security hierarchy: https://wikitech.wikimedia.org/wiki/UID

Change-Id: I50a5aa3ffa102556f2efcd5ad3c2b7a046511722