Use hiera to configure udp2log endpoint for ::mediawiki
Partially revert Iaa23f0e by restoring a class parameter for ::mediawiki
that allows setting the udp2log destination host and port. A default is
given in the class that will work for production and is also mirrored in
hiera. For the beta cluster, we will be able to set the destination via
[[:wikitech:Hiera:Deployment-prep]] to point to any host we choose
without requiring puppet realm switching or other hacks.

Bug: T1295
Bug: T74275
Change-Id: I4289fd81bd081c258153e79a3df9b5bcf0c4d417