Provision HHVM source tree in /usr/src instead of /usr/local/src
When I introduced install-pkg-src, the debug symbols for HHVM referenced a
source tree in /home/joe/HHVM/hhvm. More recent packages were built with
/usr/src/hhvm instead. I previously normalized both variants to
/usr/local/src/hhvm using 'set substitute path' in a global gdbinit.
Since we're getting rid of that, the simplest way to ensure source code context
is available to gdb is to simply provision the source tree in /usr/src and thus
eliminate the need for source path substitution.

Change-Id: Id1d40d5cb12dd9f4d1ba4d969a39530fc09c6738