ldap: fix LDAP's monitoring::service CN matching
Ignore this module's series of antipatterns and fix an underlying issue:
the monitoring alert is set up to check for a CN that is the host's
FQDN, which isn't a valid assumption as there is no gurantee those two
will match (and in fact, do not match for ldap-eqiad & ldap-codfw).
Modify the alert to check instead for the certificate's subject name
which we conveniently have under the $certificate variable already.

Change-Id: I99b26723d554750169bddceda1508af856334bcd