ldap: neptunium is eqiad, virt1000 is no more
The monitoring change (see previous commit) revealead another error:
neptunium, an eqiad host, was configured with an "ldap-codfw"
certificate.

Of course, this had no effect on the actual OpenDJ configuration,
probably because OpenDJ puppet manifests are broken in some (many?)
ways. It did affect the monitoring check, though, which noticed this
discrepancy.

Fix this, and while at it also remove the virt1000 stanza since the LDAP
role was removed from this host with commit 0510711.

Change-Id: If491512e18c4180d99957552746c3a92aec6d35d