logstash: Forward syslog events for apache2 + hhvm
Forward syslog events from apache2 and hhvm processes to a logstash
server if configured to do so via hiera. Forwards the same events that
would normally be sent to a udp2log relay directly to logstash.

Using the syslog input to logstash and rsylog as the transfer agent from
the individual MediaWiki hosts moves us closer to being able to
discontinue the use of log2udp forwarding to get events into logstash.

Bug: T76119
Change-Id: I5e8ea2b9c917ced306631c8092ce3149920a28f1