contint: mw multimedia packages on labs slaves
For some reason we are missing the `pnmtojpeg` command on the contint
labs slaves although they are on the production slaves. I am assuming
some refactoring occured between the installation of the prod slave and
the labs slave.

Including mediawiki::packages::multimedia gives us the required packaged
(ex: netbpm which provides pnmtojpeg).
We can not use mediawiki::multimedia since it set some paths to be owned
by 'apache' which is not going to work when run using the
'jenkins-deploy' user.

Bug: T76661
Change-Id: I405c2321e0239c1f4bd6ea4cba525f6840a228df