contint: mediawik::multimedia on labs slaves
For some reason we are missing the `pnmtojpeg` command on the contint
labs slaves although they are on the production slaves. I am assuming
some refactoring occured between the installation of the prod slave and
the labs slave.

Including mediawiki::multimedia will give us the cron garbage collector
and the required packaged (ex: netbpm which provides pnmtojpeg).

Bug: T76661
Change-Id: I405c2321e0239c1f4bd6ea4cba525f6840a228df