autoinstall: s/--/---/ for jessie's kernel cmdline
Linux >= 3.15 added a feature where every kernel command-line option
after a "--" is ignored. This broke d-i, which had a documented behavior
of using kernel command-line options that were passed after a "--" to
both the installer and the target. Jessie beta2+ comes with 3.16, so
this resulted in console output coming into the VGA instead of the
serial port (the only option we passed after --).

Replace this with "---" which is what is currently proposed at
https://bugs.debian.org/764675. Note that this isn't fixed yet as of
beta2, so the option is currently *not* copied to the target. It does
fix the installer output, though, and looks like something that might
get merged before the release of jessie.

Change-Id: If2366cc9cac1ed3e37a0b7de7446e9dbb907da45