autoinstall: switch to autoinstall/enable=true
Right now, we have a bunch of locale/keyboard configuration options on
the kernel command-line instead of common.cfg, as these happen before
network configuration.

However, d-i has a mode for telling it to do this after the netcfg step,
precisely for preseeding: autoinstall/enable=true, also know as "auto".
Pass this to the kernel command-line instead and move the two
configuration directives to common.cfg.

While at it, drop a bunch of deprecated and/or unneeded such options as
we only need two right now: locale & keyboard layout.

This has been tested with both precise & trusty.

Change-Id: I0a23d80b94d8ed3e0ae5b64903439c8d56b6bf6e