autoinstall: switch based on distroname too
Add an include file under common.cfg to include different files
depending on /etc/default-release, i.e. the Debian/Ubuntu version.
Populate trusty with our Ubuntu-specific preseeding commands and symlink
precise to it, as they are (and probably will stay) identical.

Add a jessie.cfg that populates Debian-specific settings: so far, just
"apt-setup/non-free" and the mirror. Use http.debian.net for the mirror,
until we get a mirror of our own, which still needs a bit more work.

Finally, drop our old per-architecture detection, as a) we are amd64
exclusively nowadays, b) the only directive that x86_64.cfg had (kernel) was not
architecture-specific but distribution-specific and got moved to
trusty.cfg instead.

Change-Id: If9b4dda82b1a1b18990caa640d4dc886642e95aa