autoinstall: set get_domain=unassigned on jessie
d-i's netcfg has a bug where the DHCP domain name option gets ignored
for preseeding purposes and manual input is required. I debugged this
extensively and filled it as https://bugs.debian.org/772523.

As we run the installer with priority=high (the default), meaning that
we get prompted for >= high questions, this breaks automated installs
for Debian. The bug exists in Ubuntu as well, but Ubuntu has a custom
patch on top of netcfg that downgrades that question to priority=medium.
The reason for this change is unknown, as in a typical Ubuntu fashion,
their custom patch is undocumented in the Ubuntu source and was never
forwarded to Debian along with a rationale. It's possible it's a
ill-conceived fix for the bug above.

Besides pushing for a fix in time for jessie, we have the following
options for a workaround:
1) Set d-i to priority=critical. This may be a good idea anyway but as
this touches far more that netcfg, it's not a great workaround right
now.
2) Remove "option domain-name" from our DHCP config; d-i would resort to
the domain found in the FQDN and the bug would not manifest. This is
not a very elegant solution.
3) Bundle the isc-dhcp-client udeb into a gzipped cpio and append it to
jessie's initrd. I approached this solution but it got messy real quick
(multiple library udeb dependencies, glibc incompatibilities etc.)
4) Trick d-i by setting netcfg/get_domain to "unassigned" (or any other
value) in the kernel command-prompt, similar to how we set
netcfg/get_hostname to "unassigned" as well. This affects only the
bootstrap phase of the installer anyway, as immediately after netcfg
comes preseeding and we preseed get_domain anyway, under our subnets/
configuration.

This implements solution (4). As the bug doesn't manifest in Ubuntu (as
explained above), this currently touches only the jessie config.

Change-Id: Ibd8cb5d4dab0b8480e09178d06d4fc4cccc68577