autoinstall: preseed xkb-keymap for jessie
console-setup 1.59 introduced keyboard-configuration/xkb-keymap (and
1.89 allowed it to be preseeded). This supersedes (replaces & enhances)
keyboard-configuration/layoutcode that we already use. Jessie currently
comes with 1.114 and prompts for more keyboard layout options without
this option. Trusty comes with 1.70ubuntu8, but 1.70ubuntu2 forked from
Debian and dropped support for xkb-keymap.

Preseed this with a value of "us" and keep our layoutcode setting for
current/future Ubuntu installs.

Change-Id: Ic8bb1685c5d5b71f60549206cad0a908ba877727