Update cached article count monthly to avoid social unrest
The link-based article count is now fast to update, as it doesn't
require page parsing as it used to until some years ago. However,
it's not instant, so it sometimes gets very out of sync. Then wrong
counts are corrected and users freak out as if Orson Welles had
announced an alien invasion at the radio.

We are sick of
* users making Reedy waste precious time to run the script manually,
* clueless discussions where we're asked to prove that no, there is
  no teapot in orbit around the sun.

Wikibooks is excluded because some subdomains use comma count which
has unknown performance. Multilingual projects are excluded because
nobody there cares about article count and some are huge.

Bug: T68867
Change-Id: I13f866b565f665a00ec6ce61bb5073cf380bd0ff