ganglia: remove legacy gmond compatibility stanzas
The "ganglia-monitor" package was named "gmond" in Ubuntu < 9.10 and had
a different config file path & init script name.

Surprisingly, we still had quite the complicated manifest to account for
this and change the behavior accordingly. Drop all this and rely just on
ganglia-monitor. As Service['gmond'] is still being used across the
tree, keep a Service alias for now.

Change-Id: I7c59993ab9a019e5dce39661e72523e8ec508f66