realm: remove $gid = '500'
realm used to have the default $gid (set to 500, wikidev), as a global
variable. Since the admin module was introduced though, this should be
unused and grepping confirmed that.

Besides being superfluous, this breaks puppet runs with a newer facter,
as facter now comes with a "gid" fact that conflicts with this global
variable of ours.

Also remove a $gid assignment for ocg hosts under site.pp, that had a
comment "what is this used for?" :)

Change-Id: Ibb2c5fe6a194d145f6be11105323d2f6f4bafe8a