reprepro: add component thirdparty
Add a component "thirdparty" to hold packages that we automatically
fetch from third parties using reprepro's update mechanisms. Currently
these end up in "main" which mixes it with packages we've built
ourselves and is a bit confusing.

Switch hwraid updates to thirdparty as well (from non-free) with the
intention of deprecating it altogether eventually, as it's unlikely
we'll ever build something non-free ourselves.

Change-Id: I948cf191cfbf3f931e6287327256e9a8d416eb76