ganglia_new: switch to stock upstart script
Ganglia's packages in trusty include an upstart script, so there's no
need to ship our own. For uniformity, replace ours with the trusty one,
which unfortunately includes a switch to the pidfile's filename, so this
means that ganglia-monitor will be restared once on all ganglia_new
hosts.

Make sure to stop provisioning the upstart script on trusty boxes, as
it's better to rely on the packages doing so, in case they update in the
future.

This also adds compatibility with non-Ubuntu systems, as it does not
depend on upstart at all.

Change-Id: I56a5174db86a35647e943a69682caed07a50bac9