Initial commit of statsv
This will eventually be a separate module, but for now it can be nested in
webperf.

StatsV is a Web request -> Varnish -> VarnishKafka -> Kafka -> StatsD gateway.
It allows client-side code to report metrics via a URL endpoint that translates
query strings to statsd format and sends them to a statsd server for
aggregation.

Change-Id: I0886b46d8a4a9115ab077b48ed538ef7e9e9cbca