HHVM: on service stop, 1-in-10 odds to clear repo
HHVM does not prune from its byte code repository code for files that have been
modified or removed. This is another case of HHVM's operational behavior being
tuned to RepoAuthoritative mode.

Some of the initial HHVM app servers are beginning to fill up /run because of
this bug. They have been up for >100 days, so if we assume that the rate at
which we push new PHP code stays more or less constant, we should plan to clear
the repo on each server at least once every that many days. But because a
server starting HHVM on an empty repo is likely to be very slow for the first
30s or so of operation, we don't want to do it everywhere at once.

So, add a random chance for clearing the bytecode repo when the service is
stopped, on the assumption that HHVM is restarted at least once every 10 days
(probably a safe assumption -- we never stopped seeing sporadic faults with
Zend, and I doubt we'll stop seeing them with HHVM any time soon.)

Change-Id: Iefdf6d7c96e98fecec7d32a53ff06ba0d9af1a62