mediawiki: hhvm_cleanup_cache - housekeeping tool for hhvm cache
The HHVM bytecode cache repository is an sqlite3 database that does not
get purged when we install a new package version and it will eventually
grow to large dimensions. This simple script fetches the current repo
schema version from HHVM, selects all the tables not belonging to that
schema and drops them. After doing that, it vacuums the sqlite3 db in
order to reduce its disk usage

Change-Id: I76a941f156369a9693a7598377803ed1e47bc821
Signed-off-by: Giuseppe Lavagetto <glavagetto@wikimedia.org>