Basic rspec setup
Add cruft for http://rspec-puppet.com/ and a NON WORKING example to
test the os_version() function in the `wmflib` module.

Usage:

  git clone ssh://gerrit.wikimedia.org:29418/labs/private.git
  bundle install
  bundle exec rspec modules/wmflib/

Class provided by labs/private.git needs to be mocked using
let(:pre_condition), that is a bit awkard but rspec-puppet does not
support passing column separated path in :module_path.  An advantage of
using :pre_condition is that we don't have to clone labs/private.git

I have added a failling test to highlight a potential incompatibility
between rspec-puppet (1.0.1) and rspec 2.x.  The rspec internal method
used to raise an error expects a parameter wich rspec-puppet does not
pass:

  1) os_version intentionally failing fails for demo
     Failure/Error: should run.with_params('ubuntu ==precise')
                               .and_return(false)
     ArgumentError:
       wrong number of arguments (0 for 1)

The error method in rspec expects a failure_message arguments which is
not provided by rspec-puppet. rspec-puppet 2.0.0 has the same issue
though it is not released yet.

Change-Id: I1c21b98dd73bb40ffd7c52e8340e245bf83ceca8