Basic rspec setup
Adds cruft for http://rspec-puppet.com/ and a NON WORKING example to
test the os_version() function in the `wmflib` module.

Usage:

  git clone ssh://gerrit.wikimedia.org:29418/labs/private.git
  bundle install
  bundle exec rspec modules/wmflib/

It fails currently apparently due to an incompatibility between
rspec-puppet (1.0.1) and rspec 2.x.  The rspec internal method used to
raise an error expects a parameter wich rspec-puppet does not pass.

rspec-puppet 2.0.0 has the same issue, it is not in rubygems anyway but
I filled a bug for it https://github.com/rodjek/rspec-puppet/issues/242

Change-Id: I1c21b98dd73bb40ffd7c52e8340e245bf83ceca8