Kill base::platform with fire
base::platform is currently a joke: it only acts on three minor server
models, one of which we don't own anymore; it sets up a $startup_devices
that is not used anywhere in our tree; and it sets up serial console
gettys using upstart, which Ubuntu automatically does anyway. Remove it
and its backreferences.

Change-Id: Iffe7a5eebdf3f93cc01b67fb9d1687bfde5a942e