Remove class base::packages::emacs
This class just installs the emacs23 package and is only included in one
place, role::statistics. Remove the base class, as it doesn't belong
there (and shouldn't have been a class anyway) and fold into
role::statistics (for now). Statistics really needs a revamp
(misc::statistics crap and all that), so hopefully it will get picked
up when that happens.

Change-Id: I8a2a571f73ae583cb66665058adbdbff2c183022