mediawiki: remove ssh.override/nice -10 for SSH
Changing the nice level of sshd (or any property of sshd) from the
mediawiki module is just a bad practice. It also breaks overriding ssh
server settings by including ssh::server unconditionally, which is now a
parameterized class and also adds a superfluous upstart dependency.
Remove the stanza entirely; it should be readded to either the ssh or
the base module if there is a need for that.

Change-Id: Ic594062ef33d2106e5974f32bb735c6c4dd2076f