Revert "reprepro: drop architecture i386"
Because multiarch. i386 is a foreign architecture that cannot be removed
on trusty systems(?!) and even if it could, we actually *do* want it on
some systems (CI, looks like).

While we could modify our apt sources to say "deb [arch=amd64]", this
seems simpler -- and who knows, we might need i386 packages at some
point for those systems too.

This reverts commit 1300ab4ebb86a9cb39c91c80ceafbedc68d23ed0.

Change-Id: If1c7f514ec2383e2371e0070bbd2c9fa5613f88a