move mediawiki maintenance scripts to module
Delete misc/maintenance.pp because we want to get rid of the entire
misc/ if possible.
Rename it to mediawiki because all these are mw maintenance scripts.
Move it into the mediawiki module in separate files.
Adjust the inclusions in site.pp.
Move files from files/scripts and other random locations into module.

Change-Id: Ie13bd7a80d8dd850cc93103aee174c15e5de469d