grub: use augeas to modify the config
The base::grub class has accumulated a bunch of test/sed/grep execs
which are hard to read and make them do the right thing, especially with
regards to other local settings that the installer may have set.

Add an augeas definition that does everything that we currently do and
has been benchmarked to be actually faster(!). Unfortunately, it looks
like that the lens is broken on systems even as new as Ubuntu trusty and
backporting a newer augeas isn't worth the trouble for now, so keep the
shell stuff as older, put behind an os_version() conditional.

Change-Id: I889d095a92bdb511f315e2b76af01da7152bfae8